Fixed warnings

+++ detail/posix/system_time.cpp
@@ -17,8 +17,8 @@ namespace pretty_output { namespace detail
 		int retval = gettimeofday(&time_value, NULL);
 		assert(retval == 0);
 
-		uint64_t microseconds = time_value.tv_usec;
-		uint64_t seconds = time_value.tv_sec;
+		uint64_t microseconds = static_cast<uint64_t>(time_value.tv_usec);
+		uint64_t seconds = static_cast<uint64_t>(time_value.tv_sec);
 		uint64_t milliseconds = seconds * 1000 + microseconds / 1000;
 
 		return milliseconds;
